temporary(?) manual http
Tor works.

This manual http request might be the only type we have to make. We are normally only going to download the same file every time, as far as I envision now.

Parsing the response with Apache should be easier than sending the request manually. At least it has message parsers.