minimum HTTP parsing
Apache does not make it straightforward to fulfill the totally normal task of taking a string and parsing it into an HTTP response. This would seem to me to be one of the most common things someone would want to do with an HTTP library. But thanks to someone on StackOverflow, we can see roughly how to do it, with the use of all kinds of extraneous objects that we shouldn't have to care about. What a mess. It works, though.