mostly JSON testing
parsing a JSON array of alerts into Alert objects, putting them into the database, and taking them back out seems ok.

fixed a race condition on adding stuff to the database. the database itself is surely always consistent and will give you what's in it, but if we're putting stuff in, we might like the rest of the program to wait before taking it back out immediately. "doesn't this remove the benefit of threading?" no, it's likely not the most common case.