notifications on alert or download error
notifications basically work, though the activities that they lead do don't do anything yet. the error notification should cause the error message to be displayed in a dialog, i think, above the main activity. the actual alert should lead to a display of that alert, with a map and so forth.